--- validate/hello/check.json
@@ -1,56 +1,0 @@
-{
-    "slug": "cs50/problems/2020/x/hello",
-    "results": [
-        {
-            "name": "exists",
-            "description": "hello.c exists",
-            "passed": true,
-            "log": [
-                "checking that hello.c exists..."
-            ],
-            "cause": null,
-            "data": {},
-            "dependency": null
-        },
-        {
-            "name": "compiles",
-            "description": "hello.c compiles",
-            "passed": false,
-            "log": [
-                "running clang hello.c -o hello -std=c11 -ggdb -lm -lcs50...",
-                "/usr/bin/../lib/gcc/x86_64-linux-gnu/7.5.0/../../../x86_64-linux-gnu/crt1.o: In function `_start':",
-                "(.text+0x20): undefined reference to `main'",
-                "clang-7: error: linker command failed with exit code 1 (use -v to see invocation)"
-            ],
-            "cause": {
-                "rationale": "code failed to compile",
-                "help": null
-            },
-            "data": {},
-            "dependency": "exists"
-        },
-        {
-            "name": "emma",
-            "description": "responds to name Emma",
-            "passed": null,
-            "log": [],
-            "cause": {
-                "rationale": "can't check until a frown turns upside down"
-            },
-            "data": {},
-            "dependency": "compiles"
-        },
-        {
-            "name": "rodrigo",
-            "description": "responds to name Rodrigo",
-            "passed": null,
-            "log": [],
-            "cause": {
-                "rationale": "can't check until a frown turns upside down"
-            },
-            "data": {},
-            "dependency": "compiles"
-        }
-    ],
-    "version": "3.1.2"
-}
