--- validate/hello/style.json
@@ -1,10 +1,0 @@
-{
-    "files": [
-        {
-            "name": "/home/ubuntu/cs50student2/hello.c",
-            "error": "file is empty"
-        }
-    ],
-    "score": 0.0,
-    "version": "2.7.4"
-}
